Refactoring - create kernels/constitutive_model.py
